change parse int to parse float for better graphing.

+++ libs/devices.js
@@ -67,7 +67,7 @@ var Device = function(id) {
         var sensor = this.getSensorById(sensorId);
         if (sensor !== null) {
             for (var i = 0; i < sensor._data.length; i++) {
-                sensorData.push(parseInt(sensor._data[i].value));
+                sensorData.push(parseFloat(sensor._data[i].value));
             }
         }
         return sensorData;
